network, pod configurators: move to new pkg
This commit also cleans up the now unnecessary attributes of
the following bind mechanisms:
  - bridge
  - masquerade
  - macvtap
