+++ pkg/virt-launcher/virtwrap/network/BUILD.bazel
@@ -9,19 +9,15 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/network",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/network:go_default_library",
         "//pkg/network/cache:go_default_library",
-        "//pkg/network/consts:go_default_library",
         "//pkg/network/dhcp:go_default_library",
         "//pkg/network/driver:go_default_library",
         "//pkg/network/errors:go_default_library",
+        "//pkg/network/infraconfigurators:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
-        "//pkg/virt-launcher/virtwrap/converter:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//staging/src/kubevirt.io/client-go/precond:go_default_library",
-        "//vendor/github.com/coreos/go-iptables/iptables:go_default_library",
-        "//vendor/github.com/vishvananda/netlink:go_default_library",
     ],
 )
 
@@ -41,6 +37,7 @@ go_test(
         "//pkg/network/dhcp:go_default_library",
         "//pkg/network/driver:go_default_library",
         "//pkg/network/errors:go_default_library",
+        "//pkg/network/infraconfigurators:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
