+++ tests/network/BUILD.bazel
@@ -21,10 +21,10 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/network/consts:go_default_library",
+        "//pkg/network/infraconfigurators:go_default_library",
         "//pkg/virt-config:go_default_library",
         "//pkg/virt-controller/services:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
-        "//pkg/virt-launcher/virtwrap/network:go_default_library",
         "//pkg/virtctl/expose:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
